Time: 43 ms (92.45%), Space: 19 MB (37.62%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.cpp
@@ -12,7 +12,7 @@ public:
     int minEatingSpeed(vector<int>& piles, int h) {
         int low = 1, high = *max_element(piles.begin(), piles.end());
         while(low <= high) {
-            long long int mid = low + (high - low) / 2;
+            int mid = low + (high - low) / 2;
             double totalHours = cal(piles, mid);
             if(totalHours <= h) {
                 high = mid - 1;
